Master view controller and data loading

+++ GoodMorning/GoodMorning/Modules/Weather/Forecast.h
@@ -13,18 +13,18 @@
 @property (weak, nonatomic) NSDate *time;
 @property (weak, nonatomic) NSString *summary;
 @property (weak, nonatomic) NSString *icon;
-@property (nonatomic) float precipIntensity;
-@property (nonatomic) float precipProbability;
-@property (nonatomic) float temperature;
-@property (nonatomic) float apparentTemperature;
-@property (nonatomic) float dewPoint;
-@property (nonatomic) float windSpeed;
-@property (nonatomic) float windBearing;
-@property (nonatomic) float cloudCover;
-@property (nonatomic) float humidity;
-@property (nonatomic) float pressure;
-@property (nonatomic) float floatfloatVisibility;
-@property (nonatomic) float floatozone;
+@property (nonatomic) double precipIntensity;
+@property (nonatomic) double precipProbability;
+@property (nonatomic) double temperature;
+@property (nonatomic) double apparentTemperature;
+@property (nonatomic) double dewPoint;
+@property (nonatomic) double windSpeed;
+@property (nonatomic) double windBearing;
+@property (nonatomic) double cloudCover;
+@property (nonatomic) double humidity;
+@property (nonatomic) double pressure;
+@property (nonatomic) double floatfloatVisibility;
+@property (nonatomic) double floatozone;
 
 - (id)initWithDictionary:(NSDictionary *)data;
